basketBallLeague/requirement.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
Faker==8.13.2
pep517==0.11.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
